refactored blackjack game logic to its own folder
