+++ frontend/src/components/LeaveModal/index.js
@@ -41,12 +41,10 @@ function LeaveModal() {
  
     
     if(type && type==='leaveTableViaTab'){
-      console.log('yes');
       dispatch(leaveTableAction(tableId))
       socket.emit('leave_table', seatObj)
 
     } else {
-      console.log('hurr');
       socket.emit('leave_seat', seatObj)
     }
