+++ frontend/src/components/TableSeat/index.js
@@ -148,7 +148,7 @@ useEffect(()=>{
     }
 
     let countdownInterval = null;
-    let countdownRemaining = Math.ceil((currentTables[activeTable.id].actionEnd - Date.now()) / 1000);
+    let countdownRemaining = Math.ceil((currentTables[activeTable.id].actionEndTimeStamp - Date.now()) / 1000);
 
     let currUser = currentTables[activeTable.id]?.tableUsers?.[seatNumber]
